+++ src/hako/utils/hako_share/impl/posix/hako_shared_memory_shm.cpp
@@ -17,24 +17,24 @@ static hako::utils::SharedMemoryInfoType* find_memory_info(
 }
 }
 
-int32_t hako::utils::HakoSharedMemoryShm::create_memory(int32_t key, int32_t size)
+int32_t hako::utils::HakoSharedMemoryShm::create_memory(int32_t key, size_t size)
 {
-    int32_t total_size = size + sizeof(SharedMemoryMetaDataType);
+    size_t total_size = size + sizeof(SharedMemoryMetaDataType);
     int32_t shmid = shmget(key, total_size, IPC_CREAT | S_IRUSR | S_IWUSR);
     if (shmid < 0) {
-        printf("ERROR: shmget() id=%d size=%d error=%d\n", key, size, errno);
+        printf("ERROR: shmget() id=%d size=%zu error=%d\n", key, size, errno);
         return -1;
     }
-    printf("INFO: shmget() key=%d size=%d \n", key, size);
+    printf("INFO: shmget() key=%d size=%zu \n", key, size);
     void *shared_memory = shmat(shmid, 0, 0);
     if (shared_memory == ((void*)-1)) {
-        printf("ERROR: shmat() id=%d size=%d error=%d\n", key, size, errno);
+        printf("ERROR: shmat() id=%d size=%zu error=%d\n", key, size, errno);
         return -1;
     }
 
     int32_t sem_id = hako::utils::sem::create(key);
     if (sem_id < 0) {
-        printf("ERROR: hako::utils::sem::create() id=%d size=%d error=%d\n", key, size, errno);
+        printf("ERROR: hako::utils::sem::create() id=%d size=%zu error=%d\n", key, size, errno);
         (void)shmdt(shared_memory);
         (void)shmctl (shmid, IPC_RMID, 0);
         return -1;
@@ -52,12 +52,12 @@ int32_t hako::utils::HakoSharedMemoryShm::create_memory(int32_t key, int32_t siz
     this->shared_memory_map_.insert(std::make_pair(key, info));
     return shmid;
 }
-void* hako::utils::HakoSharedMemoryShm::load_memory(int32_t key, int32_t size)
+void* hako::utils::HakoSharedMemoryShm::load_memory(int32_t key, size_t size)
 {
-    int32_t total_size = size + sizeof(SharedMemoryMetaDataType);
+    size_t total_size = size + sizeof(SharedMemoryMetaDataType);
     int32_t shmid = shmget(key, total_size, S_IRUSR | S_IWUSR);
     if (shmid < 0) {
-        printf("ERROR: shmget() key=%d size=%d error=%d\n", key, size, errno);
+        printf("ERROR: shmget() key=%d size=%zu error=%d\n", key, size, errno);
         return nullptr;
     }
     return this->load_memory_shmid(key, shmid);
